pyproject.toml
[project]
name = "gcpath"
version = "0.2.3"
description = "A CLI utility to manage Google Cloud Platform resource hierarchy paths."
readme = "README.md"
authors = [
    { name = "Levente", email = "leventetsk@proton.me" }
]
keywords = ["gcp", "google-cloud", "hierarchy", "folders", "projects"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
requires-python = ">=3.12"
dependencies = [
    "google-cloud-asset>=4.1.0",
    "google-cloud-resource-manager>=1.15.0",
    "rich>=14.2.0",
    "typer>=0.20.1",
]

[project.urls]
Homepage = "https://github.com/tardigrde/gcpath"
Repository = "https://github.com/tardigrde/gcpath"
Issues = "https://github.com/tardigrde/gcpath/issues"

[project.scripts]
gcpath = "gcpath.cli:run"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gcpath"]

[tool.semantic_release]
major_on_zero = false
allow_zero_version = true
version_toml = ["pyproject.toml:project.version"]
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "uv build"
commit_message = "chore(release): {version}"

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
